refactor: 边框10 tsx 改造
