+++ packages/docs/docs/Border/BorderBox10/demo.vue
@@ -1,9 +1,13 @@
 <template>
   <div w50rem h18rem p3 lex="~ " justify-center items-center bg-dark>
-    <dv-border-box-10>
+    <dv-border-box10>
       <div dv-bg>
         dv-border-box-10
       </div>
-    </dv-border-box-10>
+    </dv-border-box10>
   </div>
 </template>
+
+<script setup lang="ts">
+import { BorderBox10 as DvBorderBox10 } from '@kjgl77/datav-vue3'
+</script>
